pyproject.toml
[tool.poetry]
name = "backend-test-"
version = "0.1.0"
description = "Eccommerce backend web services"
authors = ["Ayobami Alaran A. <ayobamidele006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "4.2.11"
django-cors-headers = "4.3.1"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
python-dotenv = "1.0.1"
psycopg2-binary = "2.9.9"
django-phonenumber-field = "7.3.0"
phonenumbers = "8.13.35"


[tool.poetry.dev-dependencies]
black = "24.4.0"
flake8 = "7.0.0"
pytest = "8.1.1"
pytest-django = "4.8.0"
pytest-mock = "3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
